Added compiled tracy profiler (the viewer program) to the project.
I did this because the one I can download from the tracy repo is the wrong version, it is outdated. So I compiled it myself and added it here. I'm not sure how to set up the sub-project with the server as a proper subproject in my project, so this is the best I can do for now.